Added new actions system and "on-actions-setup" hook
The new actions system in Birdhouse provides a streamlined method for executing functions after the HTML of loaded components is added and add event listeners. This eliminates the need for timeouts in event handlers and functions, and allows the on-component-loaded, on-content-loaded, and on-actions-setup hooks to be awaited asynchronously.

When updating to this version of Birdhouse, it's recommended to migrate all event handlers to the new actions system to prevent race conditions.

The root_EXAMPLE/src/component/example.js and root_EXAMPLE/everywhere.js files have been updated to demonstrate these changes.